requirements/base.txt
Django>=2.1.2
psycopg2
Pillow